Move Access Requests guides to Access Controls
Fixes #13696

Most guides to access requests are currently not listed on the sidebar,
and the guides that are listed are available for Cloud as well as
Enterprise. This change makes it easier to find our access requests
guides by moving them to the Access Controls section and listing all
guides on the sidebar.

Since the Access Controls section includes one subsection, called
"Guides", this change also renames that subsection--and adds a
descriptive intro paragraph to the section's menu page--in order to
accommodate the Access Requests subsection.